requirements.txt
pyyaml
numpy
pydantic
opencv-python
scipy
albumentations
tqdm
Pillow


